Use service instead of repository for Products and Users

+++ src/main/java/com/vaadin/template/orders/ui/view/orders/ProductInfo.java
@@ -59,7 +59,7 @@ public class ProductInfo extends ProductInfoDesign {
 		viewEventBus.publish(new ProductInfoChange());
 	}
 
-	public double getSum() {
+	public int getSum() {
 		OrderItem item = binder.getBean();
 		return item.getQuantity() * item.getProduct().getPrice();
 	}
